% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNEMO.R
\name{affinityMatrix}
\alias{affinityMatrix}
\title{affinityMatrix}
\usage{
affinityMatrix(Diff, K = 20, sigma = 0.5)
}
\arguments{
\item{Diff}{Distance Matrix.}

\item{K}{Number of nearest neighbors.}

\item{sigma}{Variance for local model.}
}
\description{
affinityMatrix
}
\author{
Bo Wang, Aziz Mezlini, Feyyaz Demir, Marc Fiume, Zhuowen Tu, Michael Brudno, Benjamin Haibe-Kains, Anna Goldenberg
}
\keyword{internal}
