% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRRophetic.R
\name{classifySamples}
\alias{classifySamples}
\title{Dichotimize a training expression set and fit a logistic ridge regression model which is applied to the test expression matirx.}
\usage{
classifySamples(
  trainingExprData,
  trainingPtype,
  testExprData,
  batchCorrect = "eb",
  minNumSamples = 10,
  selection = -1,
  printOutput = TRUE,
  numGenesSelected = 1000,
  numSens = 15,
  numRes = 55
)
}
\arguments{
\item{trainingExprData}{Gene expression matrix for samples for which we the phenotype is already known.}

\item{trainingPtype}{The known phenotype, a vector in the same order as the columns of "trainingExprData" or with the same names as colnames of "trainingExprData".}

\item{testExprData}{Gene expression matrix for samples on which we wish to predict a phenotype. Gene names as rows, samples names as columns.}

\item{batchCorrect}{The type of batch correction to be used. Options are "eb", "none", .....}

\item{minNumSamples}{The minimum number of test samples, print an error if the number of columns of "testExprData" is below this threshold. A large number of test samples may be necessary to correct for batch effects.}

\item{selection}{How should duplicate gene ids be handled. Default is -1 which asks the user. 1 to summarize by their or 2 to disguard all duplicates.}

\item{printOutput}{Set to FALSE to supress output}

\item{numGenesSelected}{Specifies how genes are selected for "variableSelectionMethod". Options are "tTests", "pearson" and "spearman".}

\item{numSens}{The number of sensitive cell lines to be fit in the logistic regression model.}

\item{numRes}{The number of resistant cell lines fit in the logistic regression model.}
}
\value{
classifySamples
}
\description{
Dichotimize a training expression set and fit a logistic ridge regression model which is applied to the test expression matirx.
This function will return a set of probabilities.
}
\author{
Paul Geeleher, Nancy Cox, R. Stephanie Huang
}
\keyword{internal}
