% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compAgree.R
\name{compAgree}
\alias{compAgree}
\title{Comparison of agreement between two subtypes}
\usage{
compAgree(
  moic.res = NULL,
  subt2comp = NULL,
  doPlot = TRUE,
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  box.width = 0.1,
  fig.name = NULL,
  fig.path = getwd(),
  width = 6,
  height = 5
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{subt2comp}{A data.frame of subtypes that need to compare with current subtype with rownames for samples and columns for other subtypes.}

\item{doPlot}{A logic value to indicate if generating alluvial diagram to show the agreement of different subtypes; TRUE by default.}

\item{clust.col}{A string vector storing colors for each cluster.}

\item{box.width}{A numeric valur to indicate the width for box in alluvial diagram.}

\item{fig.name}{A string value to indicate the name of the figure.}

\item{fig.path}{A string value to indicate the output path for storing the figure.}

\item{width}{A numeric value to indicate the width of alluvial diagram.}

\item{height}{A numeric value to indicate the height of alluvial diagram.}
}
\value{
A figure of agreement (.pdf) if \code{doPlot = TRUE} and a data.frame storing four agreement measurements, including Rand Index (RI), Adjusted Mutual Information (AMI), Jaccard Index (JI), and Fowlkes-Mallows (FM).
}
\description{
Compute the Rand Index, Jaccard Index, Fowlkes-Mallows, and Normalized Mutual Information for agreement of two partitions, and generate alluvial diagrams for visualization.
}
\examples{
# There is no example and please refer to vignette.
}
