% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compFGA.R
\name{compFGA}
\alias{compFGA}
\title{Comparison of fraction genome altered}
\usage{
compFGA(
  moic.res = NULL,
  segment = NULL,
  iscopynumber = FALSE,
  cnathreshold = 0.2,
  test.method = "nonparametric",
  barcolor = c("#008B8A", "#F2042C", "#21498D"),
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  fig.path = getwd(),
  fig.name = NULL,
  width = 8,
  height = 4
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{segment}{A data frame containing segmented copy number and columns must exactly include the following elements: c('sample','chrom','start','end','value'). Column of `value` should be segments value when \code{iscopynumber = FALSE} but copy-number value when \code{iscopynumber = TRUE}. Copy-number will be converted to segments by log2(copy-number/2).}

\item{iscopynumber}{A logical value to indicate if the fifth column of segment input is copy-number. If segment file derived from CNV calling provides copy number instead of segment_mean value, this argument must be switched to TRUE. FALSE by default.}

\item{cnathreshold}{A numeric value to indicate the cutoff for identifying copy-number gain or loss. 0.2 by default.}

\item{test.method}{A string value to indicate the method for statistical testing. Allowed values contain c('nonparametric', 'parametric'); nonparametric means two-sample wilcoxon rank sum test for two subtypes and Kruskal-Wallis rank sum test for multiple subtypes; parametric means two-sample t-test when only two subtypes are identified, and anova for multiple subtypes comparison; "nonparametric" by default.}

\item{barcolor}{A string vector to indicate the mapping color for bars of FGA, FGG and FGL.}

\item{clust.col}{A string vector storing colors for each subtype.}

\item{fig.path}{A string value to indicate the output path for storing the barplot.}

\item{fig.name}{A string value to indicate the name of the barplot.}

\item{width}{A numeric value to indicate the width of barplot.}

\item{height}{A numeric value to indicate the height of barplot.}
}
\value{
A list contains the following components:

        \code{summary}           a table summarizing the measurements of FGA, FGG, and FGL per sample

        \code{FGA.p.value}       a nominal p value quantifying the difference of FGA among current subtypes

        \code{pairwise.FGA.test} a pairwise BH adjusted p value matrix for multiple comparisons of FGA if more than 2 subtypes were identified

        \code{FGG.p.value}       a nominal p value quantifying the difference of FGG among current subtypes

        \code{pairwise.FGG.test} a pairwise BH adjusted p value matrix for multiple comparisons of FGG if more than 2 subtypes were identified

        \code{FGL.p.value}       a nominal p value quantifying the difference of FGL among current subtypes

        \code{pairwise.FGL.test} a pairwise BH adjusted p value matrix for multiple comparisons of FGL if more than 2 subtypes were identified

        \code{test.method}       a string value indicating the statistical testing method to calculate p values
}
\description{
This function calculates Fraction Genome Altered (FGA), Fraction Genome Gained (FGG), and Fraction Genome Lost (FGL) seperately, and compares them among curent subtypes identified from multi-omics integrative clustering algorithms.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Cerami E, Gao J, Dogrusoz U, et al. (2012). The cBio Cancer Genomics Portal: An Open Platform for Exploring Multidimensional Cancer Genomics Data. Cancer Discov, 2(5):401-404.

Gao J, Aksoy B A, Dogrusoz U, et al. (2013). Integrative analysis of complex cancer genomics and clinical profiles using the cBioPortal. Sci Signal, 6(269):pl1-pl1.
}
