% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMut.R
\name{compMut}
\alias{compMut}
\title{Comparison of mutational frequency}
\usage{
compMut(
  moic.res = NULL,
  mut.matrix = NULL,
  freq.cutoff = 0.05,
  test.method = "fisher",
  p.adj.method = "BH",
  doWord = TRUE,
  doPlot = TRUE,
  innerclust = TRUE,
  res.path = getwd(),
  tab.name = NULL,
  fig.path = getwd(),
  fig.name = NULL,
  annCol = NULL,
  annColors = NULL,
  mut.col = "#21498D",
  bg.col = "#dcddde",
  p.cutoff = 0.05,
  p.adj.cutoff = 0.05,
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  width = 8,
  height = 4
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{mut.matrix}{A binary matrix storing binary mutation data with entries of 0 and 1 only.}

\item{freq.cutoff}{A numeric value to indicate the frequency cutoff for mutation data. Specifically, only features that mutated in over than such proportion would be included in testing; 0.05 by default.}

\item{test.method}{A string value to indicate statistical method for independency testing. Allowed values contain c('fisher', 'chisq'); fisher by default.}

\item{p.adj.method}{A string value to indicate the correction method for multiple comparision. Allowed values contain c('holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', 'fdr'); BH by default.}

\item{doWord}{A logic value to indicate if transformating the .txt outfile to a .docx WORD file (.txt file will be also kept); TRUE by default.}

\item{doPlot}{A logic value to indicate if generating oncoprint; TRUE by default.}

\item{innerclust}{A logic value to indicate if perform clustering within each subtype; TRUE by default.}

\item{res.path}{A string value to indicate the path for saving the table.}

\item{tab.name}{A string value to indicate the name of the output table.}

\item{fig.path}{A string value to indicate the output path for storing the oncoprint.}

\item{fig.name}{A string value to indicate the name of the oncoprint.}

\item{annCol}{A data.frame storing annotation information for samples.}

\item{annColors}{A list of string vectors for colors matched with annCol.}

\item{mut.col}{A string vector to indicate the mutation color for oncoprint.}

\item{bg.col}{A string vector to indicate the background color for oncoprint.}

\item{p.cutoff}{A numeric value to indicate the nominal p value cutoff for significant mutations shown in the oncoprint; 0.05 by default.}

\item{p.adj.cutoff}{A numeric value to indicate the adjusted p value cutoff for significant mutations shown in the oncoprint; 0.05 by default.}

\item{clust.col}{A string vector storing colors for annotating each subtype.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}
}
\value{
A figure of mutational oncoprint (.pdf) if \code{doPlot = TRUE}, a data.frame storing the difference of mutational frequency among different subtypes and a corresponding table in WORD format if \code{doWord = TRUE}.
}
\description{
This function is used to compare mutational frequency among different multi-omics integerative clusters to test the independency between subtypes and mutational status. An oncoprint will be also generated with significant mutations.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics, 32(18):2847–2849.
}
