% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compTMB.R
\name{compTMB}
\alias{compTMB}
\title{Comparsion of total mutation burden}
\usage{
compTMB(
  moic.res = NULL,
  maf = NULL,
  rmDup = TRUE,
  rmFLAGS = FALSE,
  nonSyn = NULL,
  exome.size = 38,
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  test.method = "nonparametric",
  show.size = TRUE,
  fig.path = getwd(),
  fig.name = NULL,
  width = 6,
  height = 6
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{maf}{A data frame of MAF file that has been already read with at least 10 columns as following: c('Tumor_Sample_Barcode', 'Hugo_Symbol', 'Chromosome', 'Start_Position', 'End_Position', 'Variant_Classification', 'Variant_Type', 'Reference_Allele', 'Tumor_Seq_Allele1', 'Tumor_Seq_Allele2')}

\item{rmDup}{A logical value to indicate if removing repeated variants in a particuar sample, mapped to multiple transcripts of same Gene. TRUE by default.}

\item{rmFLAGS}{A logical value to indicate if removing possible FLAGS. These FLAGS genes are often non-pathogenic and passengers, but are frequently mutated in most of the public exome studies, some of which are fishy. Examples of such genes include TTN, MUC16, etc. FALSE by default.}

\item{nonSyn}{A string vector to indicate a list of variant claccifications that should be considered as non-synonymous and the rest will be considered synonymous (silent) variants. Default value of NULL uses Variant Classifications with High/Moderate variant consequences, including c('Frame_Shift_Del', 'Frame_Shift_Ins', 'Splice_Site', 'Translation_Start_Site', 'Nonsense_Mutation', 'Nonstop_Mutation', 'In_Frame_Del', 'In_Frame_Ins', 'Missense_Mutation'). See details at \url{http://asia.ensembl.org/Help/Glossary?id=535}}

\item{exome.size}{An integer value to indicate the estimation of exome size. 38 by default (see \url{https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-017-0424-2}).}

\item{clust.col}{A string vector storing colors for annotating each Subtype.}

\item{test.method}{A string value to indicate the method for statistical testing. Allowed values contain c('nonparametric', 'parametric'); nonparametric means two-sample wilcoxon rank sum test for two subtypes and Kruskal-Wallis rank sum test for multiple subtypes; parametric means two-sample t-test when only two subtypes are identified, and anova for multiple subtypes comparison; "nonparametric" by default.}

\item{show.size}{A logical value to indicate if showing the sample size within each subtype at the top of the figure. TRUE by default.}

\item{fig.path}{A string value to indicate the output path for storing the boxviolin plot.}

\item{fig.name}{A string value to indicate the name of the boxviolin plot.}

\item{width}{A numeric value to indicate the width of boxviolin plot.}

\item{height}{A numeric value to indicate the height of boxviolin plot.}
}
\value{
A figure of TMB and TiTv distribution (.pdf) and a list with the following components:

        \code{TMB.dat}       a data.frame storing the TMB per sample within each subtype.

        \code{TMB.median}    a data.frame storing the median of TMB for each subtype.

        \code{titv.dat}      a data.frame storing the fraction contributions of TiTv per sample within each subtype.

        \code{maf.nonsilent} a data.frame storing the information for non-synonymous mutations.

        \code{maf.silent}    a data.frame storing the information for synonymous mutations.

        \code{maf.FLAGS}     a data.frame storing the information for FLAGS mutations if\code{rmFLAGS = TRUE}.

        \code{FLAGS.count}   a data.frame storing the summarization per FLAGS if\code{rmFLAGS = TRUE}.
}
\description{
This function calculates Total Mutation Burden (TMB) compares them among curent subtypes identified from multi-omics integrative clustering algorithms.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Mayakonda A, Lin D, Assenov Y, Plass C, Koeffler PH (2018). Maftools: efficient and comprehensive analysis of somatic variants in cancer. Genome Res, 28(11): 1747-1756.

Shyr C, Tarailo-Graovac M, Gottlieb M, Lee JJ, van Karnebeek C, Wasserman WW. (2014). FLAGS, frequently mutated genes in public exomes. BMC Med Genomics, 7(1): 1-14.

Chalmers Z R, Connelly C F, Fabrizio D, et al. (2017). Analysis of 100,000 human cancer genomes reveals the landscape of tumor mutational burden. Genome Med, 9(1):34.
}
