% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNEMO.R
\name{dist2}
\alias{dist2}
\title{dist2}
\usage{
dist2(X, C)
}
\arguments{
\item{X}{A data matrix where each row is a different data point.}

\item{C}{A data matrix where each row is a different data point. If this matrix is the same as X, pairwise distances for all data points are computed.}
}
\value{
dist2
}
\description{
dist2
}
\references{
Wang B, Mezlini AM, Demir F, et al (2014). Similarity network fusion for aggregating data types on a genomic scale. Nat Methods, 11(3):333-337.
}
\author{
Bo Wang, Aziz Mezlini, Feyyaz Demir, Marc Fiume, Zhuowen Tu, Michael Brudno, Benjamin Haibe-Kains, Anna Goldenberg
}
\keyword{internal}
