% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRRophetic.R
\name{estimateCI.pRRopheticCv}
\alias{estimateCI.pRRopheticCv}
\title{Confidence intervals from "pRRopheticCv" object.}
\usage{
estimateCI.pRRopheticCv(cvOutData, conf = 0.95)
}
\arguments{
\item{cvOutData}{an object of class "pRRopheticCv", i.e. the outpout of the "predictionAccuracyByCv()" function}

\item{conf}{the confidence interval required, by default 95 precent confidence interval.}
}
\value{
a numeric vector containing the average upper and lower confidence interval.
}
\description{
Given an object of class "pRRopheticCv", i.e. the output of cross validation, calculate
an average confidence interval for the predictions.
}
\author{
Paul Geeleher, Nancy Cox, R. Stephanie Huang
}
\keyword{internal}
