% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRRophetic.R
\name{estimateMedianPredictionError.pRRopheticCv}
\alias{estimateMedianPredictionError.pRRopheticCv}
\title{Median prediction error from "pRRopheticCv" object.}
\usage{
estimateMedianPredictionError.pRRopheticCv(cvOutData)
}
\arguments{
\item{cvOutData}{an object of class "pRRopheticCv", i.e. the outpout of the "predictionAccuracyByCv()" function}
}
\value{
a numeric vector containing the median prediction error from the cross validation.
}
\description{
Given an object of class "pRRopheticCv", estiamte the median prediction error,
i.e. the median difference between the predicted and measured phenotype.
}
\author{
Paul Geeleher, Nancy Cox, R. Stephanie Huang
}
\keyword{internal}
