% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClustNum.R
\name{getClustNum}
\alias{getClustNum}
\title{Get estimation of optimal clustering number}
\usage{
getClustNum(
  data = NULL,
  is.binary = rep(FALSE, length(data)),
  try.N.clust = 2:8,
  center = TRUE,
  scale = TRUE,
  fig.path = getwd(),
  fig.name = "optimal_number_cluster"
)
}
\arguments{
\item{data}{List of matrices.}

\item{is.binary}{A logicial vector to indicate if the subdata is binary matrix of 0 and 1 such as mutation.}

\item{try.N.clust}{A integer vector to indicate possible choices of number of clusters.}

\item{center}{A logical value to indicate if the variables should be centered. TRUE by default.}

\item{scale}{A logical value to indicate if the variables should be scaled. FALSE by default.}

\item{fig.path}{A string value to indicate the output figure path.}

\item{fig.name}{A string value to indicate the name of the figure.}
}
\value{
A figure that helps to choose the optimal clustering number (argument of `N.clust`) for `get%algorithm_name%()` or `getMOIC()`, and a list contains the following components:

        \code{CPI}   possible cluster number identified by clustering prediction index

        \code{Gapk}  possible cluster number identified by Gap-statistics
}
\description{
This function provides two measurements (i.e., clustering prediction index [CPI] and Gap-statistics) and aims to search the optimal number for multi-omics integrative clustering. In short, the peaks reach by the red (CPI) and blue (Gap-statistics) lines should be referred to determine `N.clust`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Chalise P, Fridley BL (2017). Integrative clustering of multi-level omic data based on non-negative matrix factorization algorithm. PLoS One, 12(5):e0176278.

Tibshirani, R., Walther, G., Hastie, T. (2001). Estimating the number of data clusters via the Gap statistic. J R Stat Soc Series B Stat Methodol, 63(2):411-423.
}
