% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOCA.R
\name{getCOCA}
\alias{getCOCA}
\title{Get subtypes from COCA}
\usage{
getCOCA(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  methods = "hclust",
  distances = "euclidean"
)
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}

\item{methods}{A string vector storing the names of clustering methods to be used to cluster the observations in each subdataset.}

\item{distances}{A string vector storing the name of distances to be used in the clustering step for each subdataset.}
}
\value{
A list with the following components:

        \code{fit}        an object returned by \link[coca]{coca}.

        \code{clust.res}  a data.frame storing sample ID and corresponding clusters.

        \code{clust.dend} a dendrogram of sample clustering.

        \code{mo.method}  a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the COCA (Cluster-of-Clusters Analysis) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Hoadley KA, Yau C, Wolf DM, et al (2014). Multiplatform analysis of 12 cancer types reveals molecular classification within and across tissues of origin. Cell, 158(4):929-944.
}
