% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusClustering.R
\name{getConsensusClustering}
\alias{getConsensusClustering}
\title{Get subtypes from ConsensusClustering}
\usage{
getConsensusClustering(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  norMethod = "none",
  reps = 500,
  pItem = 0.8,
  pFeature = 0.8,
  clusterAlg = "hc",
  innerLinkage = "ward.D",
  finalLinkage = "ward.D",
  distance = "pearson",
  plot = NULL,
  writeTable = F,
  title = file.path(getwd(), "consensuscluster"),
  seed = 123456,
  verbose = F
)
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}

\item{norMethod}{A string vector indicate the normalization method for consensus clustering.}

\item{reps}{An integer value to indicate the number of subsamples.}

\item{pItem}{A numerical value to indicate the proportion of items to sample.}

\item{pFeature}{A numerical value to indicate the proportion of features to sample.}

\item{clusterAlg}{A string value to indicate the cluster algorithm.}

\item{innerLinkage}{A string value to indicate the heirachical linakge method for subsampling.}

\item{finalLinkage}{A string value to indicate the heirarchical method for consensus matrix.}

\item{distance}{A string value to indicate the distance function.}

\item{plot}{A string value to indicate the output format for heatmap.}

\item{writeTable}{A logical value to indicate if writing output and log to csv.}

\item{title}{A string value for output directory.}

\item{seed}{A numerical value to set random seed for reproducible results.}

\item{verbose}{A logical value to indicate if printing messages to the screen to indicate progress.}
}
\value{
A list with the following components:

        \code{fit}        an object returned by \link[ConsensusClusterPlus]{ConsensusClusterPlus}.

        \code{clust.res}  a data.frame storing sample ID and corresponding clusters.

        \code{clust.dend} a dendrogram of sample clustering.

        \code{mo.method}  a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the Consensus Clustering algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Monti S, Tamayo P, Mesirov J, et al (2003). Consensus Clustering: A Resampling-Based Method for Class Discovery and Visualization of Gene Expression Microarray Data. Mach Learn, 52:91-118.
}
