% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusMOIC.R
\name{getConsensusMOIC}
\alias{getConsensusMOIC}
\title{Get subtypes from consensus clustering of multiple multi-omics integrative clustering algorithms}
\usage{
getConsensusMOIC(
  moic.res.list = NULL,
  distance = "euclidean",
  linkage = "ward.D",
  mapcolor = c("#000004FF", "#56106EFF", "#BB3754FF", "#F98C0AFF", "#FCFFA4FF"),
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  showID = FALSE,
  fig.path = getwd(),
  fig.name = "consensusheatmap",
  width = 5.5,
  height = 5
)
}
\arguments{
\item{moic.res.list}{A list of object returned by `getMOIC()`.}

\item{distance}{A string value of distance measurement for hierarchical clustering; 'euclidean' by default.}

\item{linkage}{A string value of clustering method for hierarchical clustering; 'ward.D' by default.}

\item{mapcolor}{A string vector for heatmap mapping color.}

\item{clust.col}{A string vector storing colors for annotating each cluster at the top of heatmap.}

\item{showID}{A logic value to indicate if showing the sample ID; FALSE by default.}

\item{fig.path}{A string value to indicate the output path for storing the consensus heatmap.}

\item{fig.name}{A string value to indicate the name of the consensus heatmap.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}
}
\value{
A consensus heatmap  and a list contains the following components:

        \code{consensus.hm} an object returned by \link[ComplexHeatmap]{pheatmap}

        \code{similarity}   a similary matrix for pair-wise samples with entries ranging from 0 to 1

        \code{sil}          a silhouette object that can be further passed to \link[MOVICS]{getSilhouette}

        \code{clust.res}    a data.frame storing sample ID and corresponding clusters

        \code{clust.dend}   a dendrogram of sample clustering

        \code{mo.method}    a string value indicating the method used for multi-omics integrative clustering
}
\description{
Since this R package integrates 10 mainstream multi-omics clustering algorithms, we borrow the idea of consensus clustering for later integration of the clustering results derived from different algorithms, so as to improve the clustering robustness. The simplest way to run `getConsensusMOIC()` is to pass a list of object returned by `get\%algorithm_name\%()` or by `getMOIC()` with specific argument of `methodslist`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics, 32(18):2847-2849.
}
