% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getiClusterBayes.R
\name{getiClusterBayes}
\alias{getiClusterBayes}
\title{Get subtypes from iClusterBayes}
\usage{
getiClusterBayes(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  n.burnin = 18000,
  n.draw = 12000,
  prior.gamma = rep(0.5, length(data)),
  sdev = 0.05,
  thin = 3
)
}
\arguments{
\item{data}{List of matrices with maximum of 6 subdatasets.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}

\item{n.burnin}{An integer value to indicate the number of MCMC burnin.}

\item{n.draw}{An integer value to indicate the number of MCMC draw.}

\item{prior.gamma}{A numerical vector to indicate the prior probability for the indicator variable gamma of each subdataset.}

\item{sdev}{A numerical value to indicate the standard deviation of random walk proposal for the latent variable.}

\item{thin}{A numerical value to thin the MCMC chain in order to reduce autocorrelation.}
}
\value{
A list with the following components:

        \code{fit}       an object returned by \link[iClusterPlus]{iClusterBayes}.

        \code{clust.res} a data.frame storing sample ID and corresponding clusters.

        \code{feat.res}  the results of features selection process.

        \code{mo.method} a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the iClusterBayes (Integrative clustering by Bayesian latent variable model) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Mo Q, Shen R, Guo C, Vannucci M, Chan KS, Hilsenbeck SG (2018). A fully Bayesian latent variable model for integrative clustering analysis of multi-type omics data. Biostatistics, 19(1):71-86.
}
