% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntNMF.R
\name{getIntNMF}
\alias{getIntNMF}
\title{Get subtypes from IntNMF}
\usage{
getIntNMF(data = NULL, N.clust = NULL, type = rep("gaussian", length(data)))
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}
}
\value{
A list with the following components:

        \code{fit}       an object returned by \link[IntNMF]{nmf.mnnals}.

        \code{clust.res} a data.frame storing sample ID and corresponding clusters.

        \code{mo.method} a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the IntNMF (Integrative Clustering via Non-negative Matrix Factorization) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Chalise P, Fridley BL (2017). Integrative clustering of multi-level omic data based on non-negative matrix factorization algorithm. PLoS One, 12(5):e0176278.
}
