% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRAcluster.R
\name{getLRAcluster}
\alias{getLRAcluster}
\title{Get subtypes from LRAcluster}
\usage{
getLRAcluster(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  clusterAlg = "ward.D"
)
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion; 'gaussian' by default.}

\item{clusterAlg}{A string value to indicate the cluster algorithm for similarity matrix; 'ward.D' by default.}
}
\value{
A list with the following components:

        \code{fit}        an object returned by \link[LRAcluster]{LRAcluster}.

        \code{clust.res}  a data.frame storing sample ID and corresponding clusters.

        \code{clust.dend} a dendrogram of sample clustering.

        \code{mo.method}  a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the LRAcluster (Integrated cancer omics data anlsysi by low rank approximation) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction and integrative clustering of multi-omics data using low-rank approximation: application to cancer molecular classification. BMC Genomics, 16(1):1022.
}
