% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNEMO.R
\name{getNEMO}
\alias{getNEMO}
\title{Get subtypes from NEMO}
\usage{
getNEMO(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  num.neighbors = NA
)
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}

\item{num.neighbors}{The number of neighbors to use for each omic.}
}
\value{
A list with the following components:

        \code{fit}       an object returned by \link[NEMO]{nemo.clustering}.

        \code{clust.res} a data.frame storing sample ID and corresponding clusters.

        \code{mo.method} a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the NEMO (Neighborhood based multi-omics clustering) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Rappoport N, Shamir R (2019). NEMO: cancer subtyping by integration of partial multi-omic data. Bioinformatics, 35(18):3348-3356.
}
