% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPINSPlus.R
\name{getPINSPlus}
\alias{getPINSPlus}
\title{Get subtypes from PINSPlus}
\usage{
getPINSPlus(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  norMethod = "none",
  clusteringMethod = "kmeans",
  iterMin = 50,
  iterMax = 500
)
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}

\item{norMethod}{A string vector indicate the normalization method for consensus clustering.}

\item{clusteringMethod}{The name of built-in clustering algorithm that PerturbationClustering will use. Currently supported algorithm are kmeans, pam and hclust. Default value is "kmeans".}

\item{iterMin}{The minimum number of iterations. Default value is 50}

\item{iterMax}{The maximum number of iterations. Default value is 500.}
}
\value{
A list with the following components:

        \code{fit}       an object returned by \link[PINSPlus]{PerturbationClustering}.

        \code{clust.res} a data.frame storing sample ID and corresponding clusters.

        \code{mo.method} a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the PINSPlus (Perturbation Clustering for data INtegration and disease Subtyping) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Nguyen H, Shrestha S, Draghici S, Nguyen T (2019). PINSPlus: a tool for tumor subtype discovery in integrated genomic data. Bioinformatics, 35(16):2843-2846.
}
