% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSilhouette.R
\name{getSilhouette}
\alias{getSilhouette}
\title{Get silhoutte plot from consensus clustering}
\usage{
getSilhouette(
  sil = NULL,
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  fig.path = getwd(),
  fig.name = "silhoutte",
  width = 5.5,
  height = 5
)
}
\arguments{
\item{sil}{A sil object returned from `getConsensusMOIC()`.}

\item{clust.col}{A string vector storing colors for annotating each cluster.}

\item{fig.path}{A string value to indicate the output path for storing the consensus heatmap.}

\item{fig.name}{A string value to indicate the name of the consensus heatmap.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}
}
\value{
A silhoutte barplot.
}
\description{
This function visualizes the silhoutte information.
}
\examples{
# There is no example and please refer to vignette.
}
