% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNF.R
\name{getSNF}
\alias{getSNF}
\title{Get subtypes from SNF}
\usage{
getSNF(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  K = 30,
  t = 20,
  sigma = 0.5
)
}
\arguments{
\item{data}{List of matrices.}

\item{N.clust}{Number of clusters.}

\item{type}{Data type corresponding to the list of matrics, which can be gaussian, binomial or possion.}

\item{K}{An integer value to indicate the number of neighbors in K-nearest neighbors part of the algorithm.}

\item{t}{An integer value to indicate the number of interations for the diffusion process.}

\item{sigma}{A numerical value to indicate the variance for local model.}
}
\value{
A list with the following components:

        \code{fit}       an object returned by \link[SNFtool]{SNF}.

        \code{clust.res} a data.frame storing sample ID and corresponding clusters.

        \code{mo.method} a string value indicating the method used for multi-omics integrative clustering.
}
\description{
This function wraps the SNF (Similarity Network Fusion) algorithm and provides standard output for `getMoHeatmap()` and `getConsensusMOIC()`.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Wang B, Mezlini AM, Demir F, et al (2014). Similarity network fusion for aggregating data types on a genomic scale. Nat Methods, 11(3):333-337.
}
