% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRAcluster.R
\name{LRAcluster}
\alias{LRAcluster}
\title{integrated analysis of cancer omics data by low-rank approximation}
\usage{
LRAcluster(data, types, dimension = 2, names = as.character(1:length(data)))
}
\arguments{
\item{data}{a list of data matrix,please keep the columns are the same order of samples}

\item{types}{a list of data types, can be binary, gaussian, or poisson}

\item{dimension}{the reduced dimension}

\item{names}{data names}
}
\value{
A list contains the following component:

        \code{coordinate} A matrix of the coordinates of all the samples in the reduced space

        \code{potential}  ratio of explained variance
}
\description{
The LRAcluster function is the main interface of this package, it gets a list of matrix as input and outputs the coordinates of the samples in the reduced space and the explained potential.
}
\references{
Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction and integrative clustering of multi-omics data using low-rank approximation: application to cancer molecular classification. BMC Genomics, 16(1):1022.
}
\author{
Dingming Wu, Dongfang Wang
}
\keyword{internal}
