% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNEMO.R
\name{nemo.clustering}
\alias{nemo.clustering}
\title{NEMO clustering}
\usage{
nemo.clustering(omics.list, num.clusters = NULL, num.neighbors = NA)
}
\arguments{
\item{omics.list}{A list of the data to be clustered, where each an entry is a matrix of features x samples.}

\item{k}{The number of neighbors to use for each omic. It can either be a number, a list of numbers
or NA. If it is a number, this is the number of neighbors used for all omics. If this is a list,
the number of neighbors are taken for each omic from that list. If it is NA, each omic chooses the
number of neighbors to be the number of samples divided by NUM.NEIGHBORS.RATIO.}
}
\value{
A single matrix measuring similarity between the samples across all omics.
}
\description{
Performs multi-omic clustering on a datset using the NEMO algorithm.
Uses nemo.num.clusters to estimate the number of clusters.
}
\references{
Rappoport N, Shamir R (2019). NEMO: cancer subtyping by integration of partial multi-omic data. Bioinformatics, 35(18):3348-3356.
}
\author{
Nimrod Rappoport
}
\keyword{internal}
