% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRRophetic.R
\name{plot.pRRopheticCv}
\alias{plot.pRRopheticCv}
\title{Plot "pRRopheticCv" object.}
\usage{
plot.pRRopheticCv(cvOutData)
}
\arguments{
\item{cvOutData}{an object of class "pRRopheticCv", i.e. the outpout of the "predictionAccuracyByCv()" function}
}
\value{
plot.pRRopheticCv
}
\description{
Given an object of class "pRRopheticCv", plot the cross validation
predicted values against the measured values. Also plots a regression
line.
}
\author{
Paul Geeleher, Nancy Cox, R. Stephanie Huang
}
\keyword{internal}
