% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSEA.R
\name{runGSEA}
\alias{runGSEA}
\title{Run identification of unique functional pathways}
\usage{
runGSEA(
  moic.res = NULL,
  dea.method = c("deseq2", "edger", "limma"),
  norm.expr = NULL,
  prefix = NULL,
  dat.path = getwd(),
  res.path = getwd(),
  dirct = "up",
  n.path = 10,
  msigdb.path = NULL,
  nPerm = 1000,
  minGSSize = 10,
  maxGSSize = 500,
  p.cutoff = 0.05,
  p.adj.cutoff = 0.05,
  gsva.method = "gsva",
  norm.method = "mean",
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  color = NULL,
  fig.name = NULL,
  fig.path = getwd(),
  width = 15,
  height = 10
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{dea.method}{A string value to indicate the algorithm for differential expression analysis. Allowed value contains c('deseq2', 'edger', 'limma').}

\item{norm.expr}{A matrix of normalized expression data with rows for genes and columns for samples; FPKM or TPM without log2 transformation is recommended.}

\item{prefix}{A string value to indicate the prefix of differential expression file (use for searching files).}

\item{dat.path}{A string value to indicate the path for saving the files of differential expression analysis.}

\item{res.path}{A string value to indicate the path for saving the results for identifying subtype-specific functional pathways.}

\item{dirct}{A string value to indicate the direction of identifying significant pathway. Allowed values contain c('up', 'down'); `up` means up-regulated pathway, and `down` means down-regulated pathway; "up" by default.}

\item{n.path}{A integer value to indicate how many top pathways sorted by NES should be identified for each subtypes; 10 by default.}

\item{msigdb.path}{A string value to indicate ABSOULUTE PATH/NAME of MSigDB file (GMT file with gene symbols) downloaded from \url{https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp#H}.}

\item{nPerm}{A integer value to indicate the number of permutations; 1000 by default and 10000 will be better for reproducibility.}

\item{minGSSize}{A integer value to indicate minimal size of each geneSet for analyzing; 10 by default.}

\item{maxGSSize}{A integer value to indicate maximal size of each geneSet for analyzing; 500 by default.}

\item{p.cutoff}{A numeric value to indicate the nominal p value for identifying significant pathways; pvalue < 0.05 by default.}

\item{p.adj.cutoff}{A numeric value to indicate the adjusted p value for identifying significant pathways; padj < 0.05 by default.}

\item{gsva.method}{A string value to indicate the method to employ in the estimation of gene-set enrichment scores per sample. By default this is set to gsva (Hänzelmann et al, 2013) and other options are ssgsea (Barbie et al, 2009), zscore (Lee et al, 2008) or plage (Tomfohr et al, 2005). The latter two standardize first expression profiles into z-scores over the samples and, in the case of zscore, it combines them together as their sum divided by the square-root of the size of the gene set, while in the case of plage they are used to calculate the singular value decomposition (SVD) over the genes in the gene set and use the coefficients of the first right-singular vector as pathway activity profile.}

\item{norm.method}{A string value to indicate how to calculate subtype-specific pathway enrichment scores. Allowed values contain c('mean', 'median'); mean by default.}

\item{clust.col}{A string vector storing colors for annotating each subtype at the top of heatmap.}

\item{color}{A string vector storing colors for heatmap.}

\item{fig.name}{A string value to indicate the name of the pathway heatmap.}

\item{fig.path}{A string value to indicate the output path for storing the pathway heatmap.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}
}
\value{
A figure of subtype-specific pathway heatmap (.pdf) and a list with the following components:

        \code{gsea.list}  a list storing gsea object returned by \link[clusterProfiler]{GSEA} for each subtype.

        \code{raw.es}     a data.frame storing raw enrichment score of identified subtype-specific pathways by using specified \code{gsva.method}.

        \code{scaled.es}  a data.frame storing z-scored enrichment score of identified subtype-specific pathways by using specified \code{gsva.method}.

        \code{grouped.es} a data.frame storing grouped enrichment score (mean or median value among each subtype) by using specified \code{norm.method}.

        \code{heatmap}    a complexheatmap object.
}
\description{
Use gene set enrichment analysis to identify subtype-specific (overexpressed or downexpressed) functional pathways for each subtype identified by multi-omics clustering algorithms.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Barbie, D.A. et al. (2009). Systematic RNA interference reveals that oncogenic KRAS-driven cancers require TBK1. Nature, 462(5):108-112.

Hänzelmann, S., Castelo, R. and Guinney, J. (2013). GSVA: Gene set variation analysis for microarray and RNA-Seq data. BMC Bioinformatics, 14(1):7.

Lee, E. et al. (2008). Inferring pathway activity toward precise disease classification. PLoS Comp Biol, 4(11):e1000217.

Tomfohr, J. et al. (2005). Pathway level analysis of gene expression using singular value decomposition. BMC Bioinformatics, 6(1):1-11.

Yu G, Wang L, Han Y, He Q (2012). clusterProfiler: an R package for comparing biological themes among gene clusters. OMICS, 16(5):284-287.

Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics, 32(18):2847–2849.
}
