% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVA.R
\name{runGSVA}
\alias{runGSVA}
\title{Run gene set variation analysis}
\usage{
runGSVA(
  moic.res = NULL,
  norm.expr = NULL,
  gset.gmt.path = NULL,
  gsva.method = "gsva",
  centerFlag = TRUE,
  scaleFlag = TRUE,
  halfwidth = 1,
  annCol = NULL,
  annColors = NULL,
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  distance = "euclidean",
  linkage = "ward.D",
  show_rownames = TRUE,
  show_colnames = FALSE,
  color = c("#366A9B", "#4E98DE", "#DDDDDD", "#FBCFA7", "#F79C4A"),
  fig.path = getwd(),
  fig.name = NULL,
  width = 8,
  height = 8,
  ...
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{norm.expr}{A matrix of normalized expression data with rows for genes and columns for samples; FPKM or TPM without log2 transformation is recommended.}

\item{gset.gmt.path}{A string value to indicate ABSOULUTE PATH/NAME of gene sets of interest stored as GMT format \url{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}.}

\item{centerFlag}{A logical vector to indicate if enrichment scores should be centered; TRUE by default.}

\item{scaleFlag}{A logical vector to indicate if enrichment scores should be scaled; TRUE by default.}

\item{halfwidth}{A numeric value to assign marginal cutoff for truncating enrichment scores; 1 by default.}

\item{annCol}{A data.frame storing annotation information for samples.}

\item{annColors}{A list of string vectors for colors matched with annCol.}

\item{clust.col}{A string vector storing colors for annotating each subtype at the top of heatmap.}

\item{distance}{A string value of distance measurement for hierarchical clustering; 'euclidean' by default.}

\item{linkage}{A string value of clustering method for hierarchical clustering; 'ward.D' by default.}

\item{show_rownames}{A logic value to indicate if showing rownames (feature names) in heatmap; TRUE by default.}

\item{show_colnames}{A logic value to indicate if showing colnames (sample ID) in heatmap; FALSE by default.}

\item{color}{A string vector storing colors for heatmap.}

\item{fig.path}{A string value to indicate the output path for storing the enrichment heatmap.}

\item{fig.name}{A string value to indicate the name of the enrichment heatmap.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}

\item{...}{Additional parameters pass to \link[ComplexHeatmap]{pheatmap}.}
}
\value{
A figure of enrichment heatmap (.pdf) and a list with the following components:

        \code{gset.list}  a list storing gene sets information converted from GMT format by \link[clusterProfiler]{read.gmt}.

        \code{raw.es}     a data.frame storing raw enrichment score based on given gene sets of interest by using specified \code{gsva.method}.

        \code{scaled.es}  a data.frame storing z-scored enrichment score based on given gene sets of interest by using specified \code{gsva.method}.
}
\description{
Use gene set variation analysis to calculate enrichment score of each sample in each subtype based on given gene set list of interest.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Barbie, D.A. et al. (2009). Systematic RNA interference reveals that oncogenic KRAS-driven cancers require TBK1. Nature, 462(5):108-112.

Hänzelmann, S., Castelo, R. and Guinney, J. (2013). GSVA: Gene set variation analysis for microarray and RNA-Seq data. BMC Bioinformatics, 14(1):7.

Lee, E. et al. (2008). Inferring pathway activity toward precise disease classification. PLoS Comp Biol, 4(11):e1000217.

Tomfohr, J. et al. (2005). Pathway level analysis of gene expression using singular value decomposition. BMC Bioinformatics, 6(1):1-11.

Yu G, Wang L, Han Y, He Q (2012). clusterProfiler: an R package for comparing biological themes among gene clusters. OMICS, 16(5):284-287.
}
