% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKappa.R
\name{runKappa}
\alias{runKappa}
\title{Run consistency evaluation using Kappa statistic}
\usage{
runKappa(
  subt1 = NULL,
  subt2 = NULL,
  subt1.lab = NULL,
  subt2.lab = NULL,
  fig.path = getwd(),
  fig.name = "constheatmap",
  width = 5,
  height = 5
)
}
\arguments{
\item{subt1}{A numeric vector to indicate the first appraisement. Order should be exactly the same with subt2 for each sample.}

\item{subt2}{A numeric vector to indicate the second appraisement. Order should be exactly the same with subt1 for each sample.}

\item{subt1.lab}{A string value to indicate the label of the first subtype.}

\item{subt2.lab}{A string value to indicate the label of the second subtype.}

\item{fig.path}{A string value to indicate the output path for storing the consistency heatmap.}

\item{fig.name}{A string value to indicate the name of the consistency heatmap.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}
}
\value{
A figure of consistency heatmap (.pdf).
}
\description{
Calculate Kappa statistic to measure the consistency between two appraisements
}
\details{
This function evaluates the consistency between two appraisements that targets to the same cohort.
For example, the NTP-predicted subtype amd PAM-predicted subtype of external cohort, or the current subtype and predicted subtype of discovery cohort.
Therefore, the arguments `subt1` and `subt2` can be the `clust` column of `clust.res` derived from `getMOIC()` with one specified algorithm
or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms or `runNTP()` or `runPAM()`.
However, subtypes identified from different algorithm (i.e., `get\%algorithm_name1\%` and `get\%algorithm_name2\%`) can not be evaluated
because the subtype 1 identified from the first algorithm may not be the same subtype 1 from the second algorithm.
}
\examples{
# There is no example and please refer to vignette.
}
