% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPAM.R
\name{runPAM}
\alias{runPAM}
\title{Run partition around medoids classifier}
\usage{
runPAM(
  train.expr = NULL,
  moic.res = NULL,
  test.expr = NULL,
  gene.subset = NULL
)
}
\arguments{
\item{train.expr}{A matrix of normalized expression training data with rows for genes and columns for samples; FPKM or TPM without log2 transformation is recommended.}

\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{test.expr}{A matrix of normalized expression testing data with rows for genes and columns for samples; FPKM or TPM without log2 transformation is recommended.}

\item{gene.subset}{A string vector to indicate a subset of genes to be used.}
}
\value{
A list with the following components:

        \code{IGP}        a named numeric vector storing the in-group proportion (see \link[clusterRepro]{IGP.clusterRepro}).

        \code{clust.res}  similar to `clust.res` returned by `getMOIC()` or `get%algorithm_name%` or `getConsensusMOIC()`.

        \code{mo.method}  a string value indicating the method used for prediction.
}
\description{
Using partition around medoids (PAM) classifier to predict potential subtype label on external cohort and calculate in-group proportions (IGP) statistics.
}
\details{
This function first trains a partition around medoids (PAM) classifier in the discovery (training) cohort
 to predict the subtype for patients in the external validation (testing) cohort,
 and each sample in the validation cohort was assigned to a subtype label whose centroid had the highest Pearson correlation with the sample.
 Finally, the in-group proportion (IGP) statistic will be performed to evaluate the similarity and reproducibility of the acquired subtypes between discovery and validation cohorts.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Tibshirani R, Hastie T, Narasimhan B and Chu G (2002). Diagnosis of multiple cancer types by shrunken centroids of gene expression. Proc Natl Acad Sci, 99,6567–6572.

Kapp A V, Tibshirani R. (2007). Are clusters found in one dataset present in another dataset?. Biostatistics, 8(1):9-31.
}
