% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNEMO.R
\name{spectralClustering}
\alias{spectralClustering}
\title{Spectral clustering}
\usage{
spectralClustering(affinity, K, type = 3)
}
\description{
Spectral clustering
}
\author{
Bo Wang, Aziz Mezlini, Feyyaz Demir, Marc Fiume, Zhuowen Tu, Michael Brudno, Benjamin Haibe-Kains, Anna Goldenberg
}
\keyword{internal}
