% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRRophetic.R
\name{summarizeGenesByMean}
\alias{summarizeGenesByMean}
\title{Take an expression matrix and if duplicate genes are measured, summarize them by their means}
\usage{
summarizeGenesByMean(exprMat)
}
\arguments{
\item{exprMat}{a gene expression matrix with gene names as row ids and sample names as column ids.}
}
\value{
a gene expression matrix that does not contain duplicate gene ids
}
\description{
This function accepts two expression matrices, with gene ids as rownames() and
sample ids as colnames(). It will find all duplicate genes and summarize their
expression by their mean.
}
\author{
Paul Geeleher, Nancy Cox, R. Stephanie Huang
}
\keyword{internal}
