% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRRophetic.R
\name{summary.pRRopheticCv}
\alias{summary.pRRopheticCv}
\title{Summary of "pRRopheticCv" object.}
\usage{
summary.pRRopheticCv(cvOutData)
}
\arguments{
\item{cvOutData}{an object of class "pRRopheticCv", i.e. the outpout of the "predictionAccuracyByCv()" function}
}
\value{
summary.pRRopheticCv
}
\description{
Given an object of class "pRRopheticCv", print various metrics that
summarize the performance of the cross validataion analysis
}
\author{
Paul Geeleher, Nancy Cox, R. Stephanie Huang
}
\keyword{internal}
