% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEA.R
\name{twoclassdeseq2}
\alias{twoclassdeseq2}
\title{Run two class comparison}
\usage{
twoclassdeseq2(
  moic.res = NULL,
  countsTable = NULL,
  prefix = NULL,
  overwt = FALSE,
  sort.p = TRUE,
  verbose = TRUE,
  res.path = getwd()
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{countsTable}{A matrix of RNA-Seq raw count data with rows for genes and columns for samples.}

\item{prefix}{A string value to indicate the prefix of output file.}

\item{overwt}{A logic value to indicate if to overwrite existing results; FALSE by default.}

\item{sort.p}{A logic value to indicate if to sort adjusted p value for output table; TRUE by default.}

\item{verbose}{A logic value to indicate if to only output id, log2fc, pvalue, and padj; TRUE by default.}

\item{res.path}{A string value to indicate the path for saving the results.}
}
\value{
Several .txt files storing differential expression analysis results by DESeq2
}
\description{
Two class differential expression analysis using DESeq2 algorithm.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Love, M.I., Huber, W., Anders, S. (2014) Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2. Genome Biol, 15(12):550-558.
}
\keyword{internal}
