% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEA.R
\name{twoclassedger}
\alias{twoclassedger}
\title{Run two class comparison}
\usage{
twoclassedger(
  moic.res = NULL,
  countsTable = NULL,
  prefix = NULL,
  overwt = FALSE,
  sort.p = TRUE,
  verbose = TRUE,
  res.path = getwd()
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{countsTable}{A matrix of RNA-Seq raw count data with rows for genes and columns for samples.}

\item{prefix}{A string value to indicate the prefix of output file.}

\item{overwt}{A logic value to indicate if to overwrite existing results; FALSE by default.}

\item{sort.p}{A logic value to indicate if to sort adjusted p value for output table; TRUE by default.}

\item{verbose}{A logic value to indicate if to only output id, log2fc, pvalue, and padj; TRUE by default.}

\item{res.path}{A string value to indicate the path for saving the results.}
}
\value{
Several .txt files storing differential expression analysis results by edgeR
}
\description{
Two class differential expression analysis using edgeR algorithm.
}
\details{
twoclassedger
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Robinson MD, McCarthy DJ and Smyth GK (2010). edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics, 26(1):139-140.

McCarthy DJ, Chen Y, Smyth GK (2012). Differential expression analysis of multifactor RNA-Seq experiments with respect to biological variation. Nucleic Acids Res. 40(10):4288-4297.
}
\keyword{internal}
