% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEA.R
\name{twoclasslimma}
\alias{twoclasslimma}
\title{Run two class comparison}
\usage{
twoclasslimma(
  moic.res = NULL,
  norm.expr = NULL,
  prefix = NULL,
  overwt = FALSE,
  sort.p = TRUE,
  verbose = TRUE,
  res.path = getwd()
)
}
\arguments{
\item{moic.res}{An object returned by `getMOIC()` with one specified algorithm or `get\%algorithm_name\%` or `getConsensusMOIC()` with a list of multiple algorithms.}

\item{norm.expr}{A matrix of normalized expression data with rows for genes and columns for samples; FPKM or TPM without log2 transformation is recommended.}

\item{prefix}{A string value to indicate the prefix of output file.}

\item{overwt}{A logic value to indicate if to overwrite existing results; FALSE by default.}

\item{sort.p}{A logic value to indicate if to sort adjusted p value for output table; TRUE by default.}

\item{verbose}{A logic value to indicate if to only output id, log2fc, pvalue, and padj; TRUE by default.}

\item{res.path}{A string value to indicate the path for saving the results.}
}
\value{
Several .txt files storing differential expression analysis results by limma
}
\description{
Two class differential expression analysis using limma algorithm.
}
\details{
twoclasslimma
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Ritchie, ME, Phipson, B, Wu, D, Hu, Y, Law, CW, Shi, W, and Smyth, GK (2015). limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic Acids Res, 43(7):e47.
}
\keyword{internal}
