% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cgp2016ExprRma}
\alias{cgp2016ExprRma}
\title{GDSC 2016 gene expression data}
\format{
A matrix with 17419 rows and 1018 variables.
\describe{contains "cgp2016ExprRma" the 2016 gene expression data. Data was obtained from \url{http://www.cancerrxgene.org/gdsc1000/GDSC1000_WebResources/Data/preprocessed/Cell_line_RMA_proc_basalExp.txt.zip}. Cosmic Ids (in the column names) were mapped to cell line names using data from this file: \url{ftp://ftp.sanger.ac.uk/pub/project/cancerrxgene/releases/release-6.0/Cell_Lines_Details.xlsx}}
}
\source{
\url{https://osf.io/5xvsg/}
}
\usage{
cgp2016ExprRma
}
\description{
A matrix frame storing the GDSC 2016 gene expression data.
}
\keyword{datasets}
