% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{drugData2016}
\alias{drugData2016}
\title{GDSC 2016 drug data}
\format{
A data frame with 224510 rows and 13 variables.
\describe{contains "drugData2016" the 2016 drug IC50 data, downloaded from \url{http://www.cancerrxgene.org/translation/drug/download#ic50)}}
}
\source{
\url{https://osf.io/5xvsg/}
}
\usage{
drugData2016
}
\description{
A data frame storing the GDSC 2016 drug data.
}
\keyword{datasets}
