% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNEMO.R
\name{spectralClustering}
\alias{spectralClustering}
\title{Spectral clustering}
\usage{
spectralClustering(affinity, K, type = 3)
}
\value{
spectralClustering
}
\description{
Spectral clustering
}
\references{
Wang B, Mezlini AM, Demir F, et al (2014). Similarity network fusion for aggregating data types on a genomic scale. Nat Methods, 11(3):333-337.
}
\author{
Bo Wang, Aziz Mezlini, Feyyaz Demir, Marc Fiume, Zhuowen Tu, Michael Brudno, Benjamin Haibe-Kains, Anna Goldenberg
}
\keyword{internal}
